+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added/Changed/Deprecated/Removed/Fixed/Security: YOUR CHANGE HERE
 
 * Added: `options` argument to model level attribute. Allows disabling automatic camelCase
+* Fixed: methods with complex input arguments receives `Hash` instances instead of `GraphQL::Schema::InputObject`
 
 ## [1.0.0](2020-02-07)